+++ CHANGELOG.md
@@ -5,7 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project should strive to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Template:
+
+## [v] - date
+
+### Added
+### Changed
+### Fixed
+
+---
+
 ## [] - unreleased
+
+### Added
+
+- Add migration script in scripts/migrate. When schemas for file formats are changed make upgrades safely
+- Add schema version field in saved files.
+
+### Changed
+
+- 
+### Fixed
 
 ## [v1.1.1] - 2025-01-19